# Ezima.Web/src/app/shared/api/models/child.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Reward } from '../models/reward';
import { RewardActivity } from '../models/reward-activity';
import { RewardUsage } from '../models/reward-usage';
import { User } from '../models/user';
export interface Child {
  birthday: string;
  id?: number;
  name?: string | null;
  parents?: Array<User> | null;
  rewardActivities?: Array<RewardActivity> | null;
  rewardUsages?: Array<RewardUsage> | null;
  rewards?: Array<Reward> | null;
}