# Ezima.Web/src/app/shared/api/models/reward-usage-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface RewardUsageRequest {
  childrenIds?: Array<number> | null;
  comment?: string | null;
  minutes?: number;
  usedOn?: string | null;
}