# Ezima.Web/src/app/shared/api/models/reward.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Child } from '../models/child';
import { RewardActivity } from '../models/reward-activity';
export interface Reward {
  activity?: RewardActivity;
  children?: Array<Child> | null;
  comment?: string | null;
  id?: number;
  issuedOn?: string;
  minutes?: number;
}