# Ezima.Web/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Ezima.API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/child/{childId}": {
      "get": {
        "tags": [
          "Child"
        ],
        "parameters": [
          {
            "name": "childId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Child"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Child"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Child"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/usage": {
      "get": {
        "tags": [
          "Usage"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RewardUsage"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RewardUsage"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RewardUsage"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Usage"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RewardUsageRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RewardUsageRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RewardUsageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RewardUsage"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RewardUsage"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RewardUsage"
                }
              }
            }
          }
        }
      }
    },
    "/api/usage/{usageId}": {
      "get": {
        "tags": [
          "Usage"
        ],
        "parameters": [
          {
            "name": "usageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RewardUsage"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RewardUsage"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RewardUsage"
                }
              }
            }
          }
        }
      }
    },
    "/api/usage/child/{childId}": {
      "get": {
        "tags": [
          "Usage"
        ],
        "parameters": [
          {
            "name": "childId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RewardUsage"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RewardUsage"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RewardUsage"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/usage/child/{childId}/total": {
      "get": {
        "tags": [
          "Usage"
        ],
        "parameters": [
          {
            "name": "childId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Child": {
        "required": [
          "birthday"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "maxLength": 50,
            "minLength": 3,
            "type": "string",
            "nullable": true
          },
          "birthday": {
            "type": "string",
            "format": "date-time"
          },
          "rewards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Reward"
            },
            "nullable": true
          },
          "rewardActivities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RewardActivity"
            },
            "nullable": true
          },
          "rewardUsages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RewardUsage"
            },
            "nullable": true
          },
          "parents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Reward": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "issuedOn": {
            "type": "string",
            "format": "date-time"
          },
          "activity": {
            "$ref": "#/components/schemas/RewardActivity"
          },
          "comment": {
            "type": "string",
            "nullable": true
          },
          "minutes": {
            "type": "integer",
            "format": "int32"
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Child"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RewardActivity": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "maxLength": 50,
            "minLength": 3,
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "defaultMinutes": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "RewardUsage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "usedOn": {
            "type": "string",
            "format": "date-time"
          },
          "minutes": {
            "type": "integer",
            "format": "int32"
          },
          "comment": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RewardUsageRequest": {
        "type": "object",
        "properties": {
          "minutes": {
            "type": "integer",
            "format": "int32"
          },
          "comment": {
            "type": "string",
            "nullable": true
          },
          "usedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "childrenIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "User": {
        "required": [
          "eMail",
          "fullName",
          "username"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "oAuthId": {
            "type": "string",
            "nullable": true
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "lastActive": {
            "type": "string",
            "format": "date-time"
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "eMail": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Child"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}